add option parsing.
so that we can get variable names.  parsing them, but still need
to plumb input variable names into the printer.

A sample run is:

	$ ./program --table ../input.csv --ivar A --ivar B \
	--ivar C --ivar d --ivar E --ivar F --ivar Gee --ovar X \
	--ovar Y --ovar Z --ovar W

This requires 'input.csv' to have 7 columns for the input bits and
4 columns for the output bits.

You need to add an option for the number of header lines to
skip...